+++ penman/tree.py
@@ -3,7 +3,7 @@
 Definitions of tree structures.
 """
 
-from typing import Tuple, List, Mapping, Any
+from typing import Tuple, List, Dict, Set, Mapping, Any
 
 from penman.types import (Variable, Role)
 from penman.epigraph import Epidata
@@ -47,6 +47,36 @@ class Tree:
         """
         return _nodes(self.node)
 
+    def reset_variables(self, fmt='{prefix}{j}') -> None:
+        """
+        Recreate node variables formatted using *fmt*.
+
+        The *fmt* string can be formatted with the following values:
+
+        - ``prefix``: first alphabetic character in the node's concept
+        - ``i``: 0-based index of the current occurrence of the prefix
+        - ``j``: 1-based index starting from the second occurrence
+        """
+        varmap: Dict[Variable, Variable] = {}
+        used: Set[Variable] = set()
+        for var, edges in self.nodes():
+            if var not in varmap:
+                concept = next((tgt for role, tgt, _ in edges if role == '/'),
+                               None)
+                pre = _default_variable_prefix(concept)
+                i = 0
+                newvar = None
+                while newvar is None or newvar in used:
+                    newvar = fmt.format(
+                        prefix=pre,
+                        i=i,
+                        j='' if i == 0 else i + 1)
+                    i += 1
+                used.add(newvar)
+                varmap[var] = newvar
+
+        self.node = _map_vars(self.node, varmap)
+
 
 def _format(node, level):
     var, edges = node
@@ -75,8 +105,65 @@ def _nodes(node):
     return ns
 
 
+def _default_variable_prefix(concept: Any) -> Variable:
+    """
+    Return the variable prefix for *concept*.
+
+    If *concept* is a non-empty string, the prefix is the first
+    alphabetic character in the string, if there are any, downcased.
+    Otherwise the prefix is ``'_'``.
+
+    Examples:
+        >>> default_variable_prefix('Alphabet')
+        'a'
+        >>> default_variable_prefix('chase-01')
+        'c'
+        >>> default_variable_prefix('"string"')
+        's'
+        >>> default_variable_prefix('_predicate_n_1"')
+        'p'
+        >>> default_variable_prefix(1)
+        '_'
+        >>> default_variable_prefix(None)
+        '_'
+        >>> default_variable_prefix('')
+        '_'
+    """
+    prefix = '_'
+    if concept and isinstance(concept, str):
+        for c in concept:
+            if c.isalpha():
+                prefix = c.lower()
+                break
+    return prefix
+
+
+def _map_vars(node, varmap):
+    var, edges = node
+
+    newedges = []
+    for role, tgt, epis in edges:
+        if not is_atomic(tgt):
+            tgt = _map_vars(tgt, varmap)
+        elif role != '/' and tgt in varmap:
+            tgt = varmap[tgt]
+        newedges.append((role, tgt, epis))
+
+    return (varmap[var], newedges)
+
+
 def is_atomic(x) -> bool:
     """
     Return ``True`` if *x* is a valid atomic value.
+
+    Examples:
+        >>> is_atomic('a')
+        True
+        >>> is_atomic(None)
+        True
+        >>> is_atomic(3.14
+        True
+        >>> is_atomic(('a', [('/', 'alpha', [])]))
+        False
     """
     return x is None or isinstance(x, (str, int, float))